# api-specification.yaml
openapi: 3.0.3
info:
  title: CinemaAbyss API
  description: |
    API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.
    
    Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками.
    Архитектура системы включает монолитное приложение и выделенные микросервисы.
  version: 1.0.0
  contact:
    name: CinemaAbyss Team
    email: support@cinemaabyss.com
servers:
  - url: http://localhost:8000
    description: API Gateway (Proxy Service)
  - url: http://localhost:8080
    description: Монолитное приложение
  - url: http://localhost:8091
    description: Микросервис фильмов
  - url: http://localhost:8082
    description: Микросервис событий
...
paths:
...
  /api/movies:
    get:
      tags:
        - movies
      summary: Получение списка всех фильмов
      description: Возвращает список всех фильмов с их метаданными
      operationId: getAllMovies
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - movies
      summary: Создание нового фильма
      description: Добавляет новый фильм в систему
      operationId: createMovie
      requestBody:
        description: Данные фильма для создания
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Фильм успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор фильма
          example: 1
        title:
          type: string
          description: Название фильма
          example: "Inception"
        description:
          type: string
          description: Описание фильма
          example: "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O."
        genres:
          type: array
          description: Жанры фильма
          items:
            type: string
          example: ["Sci-Fi", "Action", "Thriller"]
        rating:
          type: number
          format: float
          description: Рейтинг фильма
          example: 8.8
      required:
        - id
        - title
        - description
        - rating
    MovieInput:
      type: object
      properties:
        title:
          type: string
          description: Название фильма
          example: "Inception"
        description:
          type: string
          description: Описание фильма
          example: "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O."
        genres:
          type: array
          description: Жанры фильма
          items:
            type: string
          example: ["Sci-Fi", "Action", "Thriller"]
        rating:
          type: number
          format: float
          description: Рейтинг фильма
          example: 8.8
      required:
        - title
        - description
        - rating
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: "Internal Server Error"
      required:
        - error